add scrolling to unlocked pokemon
